requirements.txt
gradio
requests
python-dotenv
scikit-learn
pandas
matplotlib
playwright
langchain
langchain_ollama
